refactor(ai-service): split segment_classifier — extract model + dataset
First slice of the Page-5 internal ML splits. The 996-line
segment_classifier service was bundling four concerns. This commit
extracts the two cleanest leaves; the stateful SegmentClassifierService
class itself (training + inference orchestration over shared mutable
state) stays in service.py for a follow-up — a clean class-level
split needs more design care than a file-level extraction.

Extracted:
  app/ml/segment_classifier/service.py
    ├─ FocalLoss + CNN1DModel              → app/ml/segment_classifier/model.py
    └─ StreamingSegmentDataset
       + compute_derived_features          → app/storage/datasets/segment_dataset.py

Service file imports both back at the top so internal references and
external callers see the same surface. No call sites in the codebase
needed to change.

Rationale for the bands:
  - model.py imports nothing from app — a pure leaf, testable in
    isolation, no I/O.
  - segment_dataset.py owns I/O (reads from the Zarr store via
    StreamingSegmentDataset), so it belongs in app/storage/datasets/
    next to the telemetry/transformer dataset (the latter still lives
    inside transformer/model.py — same family of extraction, future PR).

Validated:
  • CNN1DModel + FocalLoss reachable through both
    `app.ml.segment_classifier.model` and the service module
    (identity preserved)
  • compute_derived_features() smoke-test produces the expected
    `*_diff` columns
  • lint-imports still 11 kept, 0 broken
  • pytest still 24 passed
  • /health 200

The other two ML monoliths (imitate_expert/service.py 1121 LOC,
transformer/model.py 2740 LOC) follow the same pattern. Each is its
own commit so reviewers can spot-check one bounded change at a time.
